web/requirements.txt
PyJWT
flask
flask_restful
gunicorn
redis
